+++ src/modules/files/c.js
@@ -1,1 +1,2 @@
+
 console.log('Hello from c.js!');
